requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiohttp==3.9.1
aiofiles==23.2.1
Pillow==10.1.0
fal-client==0.4.0
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
pytest==7.4.3
pytest-asyncio==0.21.1

